refactored folder structure to include a screens folder that mirrors pages folder colocated component files with pages they're used in created the blog page styled the blog page
