+++ pages/_app.tsx
@@ -1,8 +1,8 @@
-import '../styles/globals.css'
+import 'styles/globals.css'
 import type { AppProps } from 'next/app'
-import { Layout } from 'components/layout/layout'
+import { Layout } from 'screens/_app/layout/layout'
 
-function MyApp({ Component, pageProps }: AppProps) {
+export function MyApp({ Component, pageProps }: AppProps) {
   return (
     <Layout>
       <Component {...pageProps} />
