+++ pages/login.tsx
@@ -1,3 +1,3 @@
-export default function LoginPage() {
-  return <h1>Login</h1>
-}
+import { LoginPage } from 'screens/login/login'
+
+export default LoginPage
