+++ screens/_app/layout/header/header.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react'
-import { Nav } from 'components/layout/nav'
 import { useScrollYPosition } from 'react-use-scroll-position'
 
-import { NavMobile } from './nav-mobile'
+import { Nav } from 'screens/_app/layout/nav/nav'
+
+import { NavMobile } from 'screens/_app/layout/nav-mobile/nav-mobile'
 
 export const Header: React.FC = () => {
   const [headerOpacity, setHeaderOpacity] = React.useState(0)
@@ -28,16 +29,13 @@ export const Header: React.FC = () => {
           className="absolute w-full h-full bg-cyan-800"
         ></div>
         <div className="flex justify-between items-stretch w-full px-4 sm:px-6 md:px-12">
-          <h1 className="flex items-center font-bold text-white first-letter:sm:text-xl md:text-2xl lg:text-6xl font-semibold tracking-wider z-10">
+          <h1 className="flex items-center  text-white first-letter:sm:text-xl md:text-2xl lg:text-6xl font-semibold tracking-wider z-10">
             GY
           </h1>
           <Nav />
           <NavMobile />
         </div>
       </div>
-
-      {/* A spacer with a height that matches the header's height.  This keeps contents from being hidden behind header*/}
-      {/* <div className="h-12 md:h-16 lg:h-20"></div> */}
     </>
   )
 }
