+++ screens/_app/layout/nav/nav.tsx
@@ -1,9 +1,13 @@
-import { Menu, MenuList, MenuButton, MenuItem } from '@reach/menu-button'
-import { IDropdown } from 'interfaces-and-types'
-import Link from 'next/link'
 import { useRouter } from 'next/router'
+import Link from 'next/link'
 
-export const NavDropdown: React.FC<IDropdown> = ({
+import { navContent } from 'screens/_app/layout/constants/constants-layout'
+
+import { Menu, MenuList, MenuButton, MenuItem } from '@reach/menu-button'
+
+import { IDropdown } from 'interfaces-and-types'
+
+const NavDropdown: React.FC<IDropdown> = ({
   dropdownTitle,
   dropdownOptions,
   mainDestination,
@@ -47,5 +51,37 @@ export const NavDropdown: React.FC<IDropdown> = ({
         </>
       )}
     </Menu>
+  )
+}
+
+export const Nav: React.FC = () => {
+  const router = useRouter()
+
+  return (
+    <div className="hidden sm:flex sm:w-104 md:w-120 lg:w-152 justify-between space-between items-center tracking-wider text-md md:text-lg lg:text-xl z-20">
+      {navContent.dropdowns.map(dropdown => {
+        return (
+          <NavDropdown
+            key={dropdown.dropdownTitle}
+            dropdownTitle={dropdown.dropdownTitle}
+            dropdownOptions={dropdown.dropdownOptions}
+            mainDestination={dropdown.mainDestination}
+          />
+        )
+      })}
+      {navContent.navLinks.map(navLink => {
+        return (
+          <Link key={navLink.linkText} href={navLink.destination}>
+            <a
+              className={` text-white hover:text-cyan-200  ${
+                router.pathname === navLink.destination ? 'text-cyan-200' : ''
+              }  `}
+            >
+              {navLink.linkText}
+            </a>
+          </Link>
+        )
+      })}
+    </div>
   )
 }
