+++ screens/consulting/consulting.tsx
@@ -1,0 +1,3 @@
+export function ConsultingPage() {
+  return <h1>Consulting</h1>
+}
