+++ screens/home/home.tsx
@@ -1,0 +1,7 @@
+export function Home() {
+  return (
+    <div>
+      <h1>GY Consulting</h1>
+    </div>
+  )
+}
