+++ screens/homestay/homestay.tsx
@@ -1,0 +1,3 @@
+export function HomestayPage() {
+  return <h1>Homestay</h1>
+}
