+++ screens/login/login.tsx
@@ -1,0 +1,3 @@
+export function LoginPage() {
+  return <h1>Login</h1>
+}
